+++ src/gui/widgets/dialogs/glaxnimate_window_p.hpp
@@ -307,6 +307,7 @@ public:
     void setup_layers_actions();
     void action_add_keyframe();
     void action_add_keyframe_for(model::Object *object, model::AnimatableBase* animatable);
+    void action_keyframe_transition(model::KeyframeTransition::Descriptive new_trans);
 
     void init_plugins();
     void init_tools_ui();