+++ CHANGELOG.md
@@ -25,6 +25,7 @@
     * Small improvements to the gradients docks (visual changes, easier to edit, and duplicate button)
     * Added "Focus Mode" which hides all controls from the canvas
     * KDE Find Action...
+    * Menu actions to set the transition of selected keyframes
 * I/O:
     * Improved matte import and export for lottie
     * Improved mask properties import for lottie